Module serenity_framework::argument

Utilities for parsing command arguments.

Structs

ParseEitherError

Error that is returned when ParseEither::parse fails.

Enums

ArgumentError

Error that might have occured when trying to parse an argument.

ParseEither

Denotes a type that can be either one of two different types.

Functions

optional_argument_from_str

Tries to take a single segment from a list of segments and parse an argument out of it using the std::str::FromStr trait.

optional_argument_parse

Tries to take a single segment from a list of segments and parse an argument out of it using the serenity::utils::Parse trait.

required_argument_from_str

Takes a single segment from a list of segments and parses an argument out of it using the std::str::FromStr trait.

required_argument_parse

Takes a single segment from a list of segments and parses an argument out of it using the serenity::utils::Parse trait.

rest_argument_from_str

Parses the remainder of the list of segments into an argument using the std::str::FromStr trait.

rest_argument_parse

Parses the remainder of the list of segments into an argument using the serenity::utils::Parse trait.

variadic_arguments_from_str

Tries to parse many arguments from a list of segments using the std::str::FromStr trait.

variadic_arguments_parse

Tries to parse many arguments from a list of segments using the serenity::utils::Parse trait.