Function serenity_framework::argument::required_argument

pub fn required_argument<T, D>(
    ctx: &Context<D>, 
    segments: &mut ArgumentSegments<'_>
) -> Result<T, ArgumentError<T::Error>> where
    T: Argument<D>, 

Takes a single segment from a list of segments and parses an argument out of it.

Errors

If the list of segments is empty, ArgumentError::Missing is returned.

If the segment cannot be parsed into an argument, ArgumentError::Argument is returned.