Trait serenity_framework::argument::Argument

pub trait Argument<D = DefaultData>: Sized {
    type Error;
    pub fn parse(ctx: &Context<D>, source: &str) -> Result<Self, Self::Error>;
}

Abstraction for parsing arguments from a source.

Associated Types

type Error

Type of error that may be returned from trying to parse a source.

Loading content...
Required methods

pub fn parse(ctx: &Context<D>, source: &str) -> Result<Self, Self::Error>

Parses a source into the argument type, with auxiliary information from a Context.

Loading content...