Module serenity_framework::command

Functions and types relating to commands.

A command is a function that performs work. It is invoked by a user on Discord. It may have many names by which it can be invoked, but will always have at least one name. It may possess subcommands to arrange functionality together. It may have information that relays to the user what it does, what it is for, and how it is used. It may have checks to allow/deny a user’s access to the command.

Structs

Command

Data surrounding a command.

CommandBuilder

A builder type for creating a Command from scratch.

CommandId

A unique identifier of a Command stored in the CommandMap.

Type Definitions

CommandConstructor

A constructor of the Command type provided by the consumer of the framework.

CommandFn

The definition of a command function.

CommandMap

IdMap for storing commands.

CommandResult

The result type of a command function.

StringHook

A function to dynamically create a string.

StringsHook

A function to dynamically create a list of strings.