Module serenity_framework::context

Data provided in different contexts.

A context type contains data that is only available in certain phases of command dispatch. Other than the Discord message, data is placed into the context types in order to arrange them together and to allow extending the types in future releases without breaking function definitions.

Structs

CheckContext

The check context.

Context

The final context type.

PrefixContext

The prefix context.