Function serenity_framework::parse::content::command

pub async fn command<'a, D, E, F, Fut, '_, '_>(
    conf: &'a Configuration<D, E>, 
    segments: &'_ mut Segments<'_>, 
    group: Option<&'a Group<D, E>>, 
    f: F
) -> Result<(&'a Group<D, E>, &'a Command<D, E>), Error<E>> where
    F: Fn(&'a Group<D, E>, &'a Command<D, E>) -> Fut,
    Fut: Future<Output = Result<(), Error<E>>>, 

Parses a Command out of the content.

The last command and the group it or its predecessor belongs to are returned.

A content can have an infinite amount of commands, so long they are subcommands of the commands preceeding them.

This function behaves differently when a group is passed or not. If the group is present, the command that will be returned may be:

from the group's commands list; or

a subcommand (or its subcommand...) of the command from the group; or

the group's default command if defined; or

a subcommand (or its subcommand...) of the default command.

If the group is not the present, only the 1. and 2. points count.

If the group is present

If the passed segments are empty, and the default command is not defined, the MissingContent error is returned. If the latter is defined, it is then chosen as the initial command.

If the group is not present

If the passed segments are empty, the MissingContent error is returned.

If the group is present

If the first segment is not a command name, and the default command is not defined, the InvalidCommandName error is returned. If the latter is defined, it is then chosen as the initial command.

If the group is not present

If the first segment is a name not affiliated with any command, the InvalidCommandName error is returned. Otherwise, the command under that name is chosen as the initial command.

If the group is not present, a search through Configuration::top_level_groups will be conducted. If this fails, the InvalidCommand error without the group id is returned.

As the function successfully parses a command, the given f async function is called with the command and its group. Its purpose is to short-circuit this function if the command does not conform to certain conditions, such as its checks failing. If f returns an error, it is propagated to the caller of this function.