Function serenity_framework::parse::content::group

pub async fn group<'a, D, E, F, Fut, '_, '_>(
    conf: &'a Configuration<D, E>, 
    segments: &'_ mut Segments<'_>, 
    f: F
) -> Result<Option<&'a Group<D, E>>, Error<E>> where
    F: Fn(&'a Group<D, E>) -> Fut,
    Fut: Future<Output = Result<(), Error<E>>>, 

Parses a Group out of the content.

If no groups are present in the content, None is returned. Otherwise, the last group is returned.

A content can have an infinite amount of groups, so long they have prefixes and are subgroups of the groups preceeding them.

A Top Level Group cannot be returned from this function, as they do not have prefixes and subgroups.

As the function successfully parses a group, the given f async function is called with the group. Its purpose is to short-circuit this function if the group does not conform to certain conditions, such as its checks failing. If f returns an error, it is propagated to the caller of this function.