Function serenity_framework::parse::commands

pub fn commands<'a, 'b, 'c, D, E>(
    conf: &'a Configuration<D, E>, 
    segments: &'b mut Segments<'c>
) -> CommandIterator<'a, 'b, 'c, D, E>

Creates a command parsing iterator.

The returned iterator will iterate through the segments of the message, returning each valid command that it can find.

Return type of the iterator

The iterator will return items of the type Result<&Command,DispatchError>.

The Result signifies whether a given name for the first command exists. If it is not the case, the InvalidCommandName error is returned.

The Option returned from calling Iterator::next will signify whether the content had a command, did not have a command, or was empty.