Module serenity_framework::utils::id_map

An Identifier Map. An abstraction for structures who may have many names, but only once instance.

The Identifier Map, or IdMap for short, handles the case when a structure is stored once, but may be retrieved using a variety of names or aliases. A naive approach would be a simple HashMap<String, Struct>. However, this is inefficient. You would have to keep copies of the structure for each of its name. To avoid this, the IdMap assigns a unique identifier to the structure that is cheap to copy and small to store in memory. Consequently, instead of keeping copies of the structure for each of its name, we do this for the identifier. The structure can then be retrieved with the identifier. The IdMap employs two HashMaps to accomplish its job. For small structures, IdMap might be inefficient memory-wise and processor-wise. But it can pay off well for big/huge structures, which may be hundreds of bytes long.

The IdMap is generic. You can use it with any type for the name of the structure, the identifier for the structure, and the structure itself.

Examples

Using IdMap for your own purposes:

use serenity_framework::utils::IdMap;

#[derive(Debug, PartialEq)]
struct Foo {
    bar: i32,
    baz: String,
}

#[derive(Clone, Copy, PartialEq, Eq, Hash)]
struct FooId(u64);

let mut map: IdMap<String, FooId, Foo> = IdMap::new();

let foo1 = Foo { bar: 1, baz: "2".to_string() };
let foo2 = Foo { bar: 3, baz: "4".to_string() };

map.insert_name("fo".to_string(), FooId(1));
map.insert_name("foo".to_string(), FooId(1));
map.insert(FooId(1), foo1);

map.insert_name("go".to_string(), FooId(2));
map.insert(FooId(2), foo2);

assert_eq!(map.get(FooId(1)), Some(&Foo { bar: 1, baz: "2".to_string() }));
// This will panic if a structure under that identifier does not exist.
assert_eq!(&map[FooId(1)], &Foo { bar: 1, baz: "2".to_string() });
assert_eq!(map.get_by_name("fo"), Some(&Foo { bar: 1, baz: "2".to_string() }));
assert_eq!(map.get_by_name("foo"), Some(&Foo { bar: 1, baz: "2".to_string() }));

assert_eq!(&map[FooId(2)], &Foo { bar: 3, baz: "4".to_string() });
assert_eq!(map.get_by_name("go"), Some(&Foo { bar: 3, baz: "4".to_string() }));
assert_eq!(map.get_by_name("goo"), None);

Structs

IdMap
An Identifier Map. An abstraction for structures who may have many names, but only once instance.