Module serenity_framework::utils::segments

Functions and types for handling segments.

A segment is a substring of a source string. The boundaries of the substring are determined by a delimiter, which is a &str value.

Structs

ArgumentSegments

An iterator type that splits a string into argument segments using a delimiter and quotes.

Segments

An iterator type that splits a string into segments using a delimiter.

Functions

argument_segment

Returns an argument segment of the source.

argument_segment_split

Returns an argument segment and the rest of the source.

quoted_segment

Returns a quoted segment of the source.

quoted_segment_split

Returns a quoted segment and the rest of the source.

segment

Returns a segment of the source.

segment_index

Returns the index to the end of a segment in the source.

segment_split

Returns a segment and the rest of the source after the delimiter.