Function serenity_framework::utils::segments::argument_segment_split

pub fn argument_segment_split<'a>(
    src: &'a str, 
    delimiter: &str
) -> Option<(&'a str, &'a str)>

Returns an argument segment and the rest of the source.

An argument segment is either a quoted segment or a normal segment.

When the segment is quoted, the rest of the source is trimmed off of the specified delimiter.

If the source is empty, None is returned.

Examples

// Used example strings are from the YouTube video https://www.youtube.com/watch?v=1edPxKqiptw
use serenity_framework::utils::argument_segment_split;

assert_eq!(argument_segment_split("", ", "), None);
assert_eq!(argument_segment_split("Font, front, wont", ", "), Some(("Font", "front, wont")));
assert_eq!(
    argument_segment_split("\"want, grand\", and grant", ", "),
    Some(("want, grand", "and grant"))
);
assert_eq!(
    argument_segment_split("\"Shoes, goes, does.", ", "),
    Some(("Shoes, goes, does.", ""))
);