Crate serenity_framework

The official command framework for Serenity bots.

The framework provides an interface between functionality of the bot and a user on Discord through the concept of commands. They are functions that the user invokes in a guild channel or private channel.

Command invocations start with a prefix at the beginning of the message. The prefix distinguishes normal messages and command invocations. If the prefix is unique, it also avoids collision with command invocations of other bots. The bot may have many prefixes, statically or dynamically defined.

Assuming the prefix is ! and a command with the name ping exists, a typical invocation might look like:

!ping

Commands can accept arguments. These are the content of the message after the command name. As an example:

!sort 4 2 8 -3

The arguments of the sort command is a "4 2 8 -3" string. Arguments are not processed by the framework, as it is the responsibility of each command to decide the correct format of its arguments, and how they should be parsed.

Commands may be categorized. A category is a list of individual commands with a common theme, such as moderation. They do not participate in command invocation. They are used to register commands in bulk and display related commands in the help command.

Modules

argument
Utilities for parsing command arguments.

category
A list of separate, but related commands.

check
Functions and types relating to checks.

command
Functions and types relating to commands.

configuration
Configuration of the framework.

context
Data provided in different contexts.

error
Defines error types used by the framework.

parse
Utilities to parse the prefix and command out of a message.

prelude
A series of re-exports to simplify usage of the framework.

utils
A set of abstraction utilities used by the framework to simplify its code.

Structs

Framework
The core of the framework.

Type Definitions

DefaultData
The default type for user data when it is unspecified.