Module serenity_framework::check

Functions and types relating to checks.

A check is a function that can be plugged into a command to allow/deny a user’s access. The check returns a Result that indicates whether it succeeded or failed. In the case of failure, additional information can be given, a reason, that describes the failure.

Structs

Check

Data relating to a check.

CheckBuilder

A builder type for creating a Check from scratch.

Enums

Reason

The reason describing why a check failed.

Type Definitions

CheckConstructor

A constructor of the Check type provided by the consumer of the framework.

CheckFn

The definition of a check function.

CheckResult

The result type of a check function