Struct serenity_framework::configuration::Configuration

#[non_exhaustive]pub struct Configuration<D = DefaultData, E = DefaultError> {
    pub prefixes: Vec<String>,
    pub dynamic_prefix: Option<DynamicPrefix<D, E>>,
    pub case_insensitive: bool,
    pub no_dm_prefix: bool,
    pub on_mention: Option<String>,
    pub categories: Vec<Category>,
    pub root_level_commands: HashSet<CommandId>,
    pub commands: CommandMap<D, E>,
}

The configuration of the framework.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct {{ .. }} syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.
prefixes: Vec<String>

A list of static prefixes.

dynamic_prefix: Option<DynamicPrefix<D, E>>

A function to dynamically parse the prefix.

case_insensitive: bool

A boolean indicating whether casing of the letters in static prefixes, or command names does not matter.

no_dm_prefix: bool

A boolean indicating whether the prefix is not necessary in direct messages.

on_mention: Option<String>

A user id of the bot that is used to compare mentions in prefix position.

If filled, this allows for invoking commands by mentioning the bot.

categories: Vec<Category>

A list of Categorys.

root_level_commands: HashSet<CommandId>

A set of commands that can only appear at the beginning of a command invocation.

commands: CommandMap<D, E>

An IdMap containing all Commands.

Implementations

impl<D, E> Configuration<D, E>

pub fn new() -> Self

Creates a new instance of the framework configuration.

pub fn prefix<I>(&mut self, prefix: I) -> &mut Self where
I: Into<String>,

Assigns a prefix to this configuration.

The prefix is added to the prefixes list.

pub fn dynamic_prefix(&mut self, prefix: DynamicPrefix<D, E>) -> &mut Self

Assigns a function to dynamically parse the prefix.

pub fn case_insensitive(&mut self, b: bool) -> &mut Self

Assigns a boolean indicating whether the casing of letters in static prefixes, or command names does not matter.

pub fn no_dm_prefix(&mut self, b: bool) -> &mut Self

Assigns a boolean indicating whether the prefix is not necessary in direct messages.

pub fn on_mention<I>(&mut self, id: I) -> &mut Self where
I: Into<UserId>,

Assigns a user id of the bot that will allow for mentions in prefix position.

pub fn category<I>(
&mut self,
name: I,
cmds: &[CommandConstructor<D, E>]
) -> &mut Self where
I: Into<String>,

Assigns a category to this configuration.

The category is added to the categories list. Additionally, all of its commands are added to the commands map

pub fn command(&mut self, command: CommandConstructor<D, E>) -> &mut Self

Assigns a command to this configuration.

The command is added to the commands map, alongside its subcommands. It it also added into the root_level_commands set.