Struct serenity_framework::utils::id_map::IdMap

pub struct IdMap<Name, Id, Struct> { /* fields omitted */ }

An Identifier Map. An abstraction for structures who may have many names, but only once instance.

Refer to the module-level documentation

Implementations

impl<Name, Id, Struct> IdMap<Name, Id, Struct>

pub fn new() -> Self

Creates a new IdMap instance.

pub fn len_names(&self) -> usize

Returns the total number of names stored.

pub fn len(&self) -> usize

Returns the total number of structures stored.

pub fn is_empty(&self) -> bool

Returns the a boolean indicating that the map is empty.

The map is regarded as empty when it contains no structures.

pub fn iter_names(&self) -> Keys<'_, Name, Id>

Returns an iterator over all names stored in the map.

pub fn iter_ids(&self) -> Values<'_, Name, Id>

Returns an iterator over all identifiers stored in the map.

Duplicate identifiers may appear.

pub fn iter(&self) -> Iter<'_, Id, Struct>

Returns an iterator over all structures and their assigned identifier.

pub fn iter_mut(&mut self) -> IterMut<'_, Id, Struct>

Returns a mutable iterator over all structures and their assigned identifier.

Only the structures are mutable.

impl<Name, Id, Struct> IdMap<Name, Id, Struct> where
Name: Hash + Eq,
Id: Hash + Eq + Copy,

pub fn insert_name(&mut self, name: Name, id: Id) -> Option<Id>

Assigns a name to an identifier.

Returns None if the name does not exist in the map.

Returns Some(old_id) if the name exists in the map. The identifier is overwritten with the new identifier.

pub fn get_id<B: ?Sized>(&self, name: &B) -> Option<Id> where
Name: Borrow<B>,
B: Hash + Eq,

Retrieves an identifier based on a name.

A copy of the identifier is returned.

Returns None if an identifier does not belong to the name, otherwise Some.

pub fn get_by_name<B: ?Sized>(&self, name: &B) -> Option<&Struct> where
Name: Borrow<B>,
B: Hash + Eq,

Retrieves a structure based on an identifier.

An immutable reference to the structure is returned.

Returns None if a structure does not belong to the name, otherwise Some.

pub fn get_by_name_mut<B: ?Sized>(&mut self, name: &B) -> Option<&mut Struct> where
Name: Borrow<B>,
B: Hash + Eq,

Retrieves a structure based on an identifier.

A mutable reference to the structure is returned.

Returns None if a structure does not belong to the name, otherwise Some.

pub fn get_pair<B: ?Sized>(&self, name: &B) -> Option<(Id, &Struct)> where
Name: Borrow<B>,
B: Hash + Eq,

Retrieves both an identifier and its structure based on a name.

An identifier and an immutable reference to the structure is returned.

Returns None if a identifier/structure does not belong to the name, otherwise Some.

pub fn contains<B: ?Sized>(&self, name: &B) -> bool where
Name: Borrow<B>,
B: Hash + Eq,

Returns a boolean indicating that a structure exists under a name

impl<Name, Id, Struct> IdMap<Name, Id, Struct> where
Id: Hash + Eq,

pub fn insert(&mut self, id: Id, aggr: Struct) -> Option<Struct>

Assigns a structure to an identifier.

Returns None if the identifier does not exist in the map.

Returns Some(old_struct) if the identifier exists in the map. The structure is overwritten with the new structure.

pub fn get(&self, id: Id) -> Option<&Struct>

Retrieves a structure based on an identifier.

An immutable reference is returned.

Returns None if a structure does not belong to the identifier, otherwise Some.

pub fn get_mut(&mut self, id: Id) -> Option<&mut Struct>

Retrieves a structure based on an identifier.

An mutable reference is returned.

Returns None if a structure does not belong to the identifier, otherwise Some.